file seperate for shape filter
